server/pyproject.toml
[project]
name = "campugrid-server"
version = "0.1.0"
description = "CampuGrid — Distributed P2P Compute Platform Server"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "websockets>=14.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    
    # Redis
    "redis[hiredis]>=5.2.0",
    
    # Task queue
    "celery[redis]>=5.4.0",
    
    # Object storage
    "minio>=7.2.12",
    
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    
    # Validation & serialization
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    
    # AI Pipeline
    "google-generativeai>=0.8.0",
    
    # File analysis
    

    # Monitoring
    "prometheus-fastapi-instrumentator>=7.0.2",
    
    # Utilities
    "python-dotenv>=1.0.1",
    "asgiref>=3.8.0",
    "sentry-sdk[fastapi]>=2.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
